backend/requirements.txt
Flask
flask-cors
opencv-python
numpy
ultralytics
python-dotenv
